+++ src/components/index.ts
@@ -2,5 +2,3 @@ export * from "./drawer-navigation"
 export * from "./order-list";
 export * from "./game-card"
 export * from "./query-container"
-export * from "./channel";
-export * from "./channel-list";
