+++ src/screens/game-detail/index.ts
@@ -1,5 +1,7 @@
-export * from "./chat-mobile";
+export * from "./channel-list";
 export * from "./channel";
+export * from "./chat";
+export * from "./create-channel";
 export * from "./map";
 export * from "./orders";
 export * from "./layout";
